reorganize code and add processing results
