+++ geospaas_rest_api/pagination.py
@@ -1,3 +1,4 @@
+""""""
 from rest_framework.pagination import CursorPagination
 
 class PKOrderedCursorPagination(CursorPagination):
@@ -8,3 +9,15 @@ class PKOrderedCursorPagination(CursorPagination):
     ordering = 'pk'
     page_size = 100
     page_size_query_param = 'page_size'
+
+
+class DateOrderedCursorPagination(CursorPagination):
+    """Pagination class ordering by decreasing date_created"""
+    ordering = '-date_created'
+    page_size = 100
+
+
+class IdOrderedCursorPagination(CursorPagination):
+    """Pagination class ordering by decreasing date_created"""
+    ordering = '-id'
+    page_size = 100
